requirements.txt
pyodbc~=4.0.35
python-dotenv~=1.0.0
pyzabbix~=1.2.1
pythonping~=1.1.4
requests~=2.28.2